Add files via upload
This project is a music playlist. In the Music Playlist you are allowed to input Song name, Artist, Genre, Duration, and Release Year, this information into a cloud that will store your info. Additionally you're able choose between either having an ascending or decending order of either of the options listed.